refactor: Extract @Transactional chunk processor and inject into ProcessDailyStatisticsService
